requirement.txt
Pillow==9.5.0
pyttsx3==2.90
speech_recognition==3.10.0
langchain_google_genai==0.1.0
langchain_core==0.1.0
pywhatkit==2.1
mss==7.0.1
pyautogui==0.9.53
opencv-python==4.7.0.72
numpy==1.25.0
psutil==5.9.4
requests==2.31.0
playsound==1.2.2
openai==0.27.1
